refactor: extract TransactionService; route broadcast through WalletProvider
