+++ src/main/src/providers/P2PWalletProvider.ts
@@ -1,4 +1,4 @@
-import {Block, Script, utils as sdkUtils} from 'dash-core-sdk'
+import {Block, Script, Transaction as SDKTransaction, utils as sdkUtils} from 'dash-core-sdk'
 import {UTXO} from '../types/UTXO'
 import {Transaction} from '../types/Transaction'
 import {TransactionDAO} from '../database/TransactionDAO'
@@ -9,13 +9,15 @@ const {addressToPublicKeyHash} = sdkUtils
 // Reads wallet info from the local SQL database populated by SPV cfilter
 // sync. Only knows about txs that touched our addresses.
 //
+// Broadcast is delegated to the `broadcastFallback` provider — native P2P
+// inv/tx broadcast through the utility process isn't implemented yet.
 // Unsupported (throw):
-//   - broadcastTx
 //   - getBlockByHash
 export class P2PWalletProvider implements WalletProvider {
   constructor(
     private readonly transactionDAO: TransactionDAO,
     private readonly walletId: string,
+    private readonly broadcastFallback: WalletProvider,
   ) {}
 
   async getTransactions(address: string): Promise<Transaction[]> {
@@ -43,8 +45,8 @@ export class P2PWalletProvider implements WalletProvider {
     }))
   }
 
-  async broadcastTx(): Promise<string> {
-    throw new Error('Unimplemented: broadcastTx is not available in p2p mode')
+  async broadcastTx(tx: SDKTransaction): Promise<string> {
+    return this.broadcastFallback.broadcastTx(tx)
   }
 
   async getBlockByHash(): Promise<Block> {
